feat: 대시보드 훅 생성
