--- frontend/src/hooks/dashboard/useDashboard.ts
@@ -1,17 +1,0 @@
-import type { CompetitionId } from '@/apis/competitions';
-
-type Rank = {
-  user: string;
-  solved: number;
-  score: number;
-  problemSet: {
-    [key: number]: number | null;
-  };
-};
-
-type Dashboard = {
-  competitionId: CompetitionId;
-  totalProblemCount: number;
-  ranking: Rank[];
-  myRanking: Rank;
-};
